UI updated; user icon added to delete member; address field added to exising members

+++ app/util/GymUtil.java
@@ -6,7 +6,7 @@ import models.Member;
 public class GymUtil {
 
     public static double calculateBMI(Member m, Assessment a){
-        double bmiValue = a.weight/(m.getHeight() * m.getHeight());
+        double bmiValue = (double) a.weight/(m.getHeight() * m.getHeight());
         return bmiValue;
     }
